Rename `EmailProviderCache` to `MessageListCache`
